+++ lunatrace/bsl/backend/src/models/vulnerability-dependency-tree/index.ts
@@ -16,7 +16,6 @@ import { randomUUID } from 'crypto';
 import { SeverityNamesOsv } from '@lunatrace/lunatrace-common/build/main';
 
 import { Analysis_Finding_Type_Enum } from '../../hasura-api/generated';
-import { newBenchmarkTable, start } from '../../utils/performance';
 import { notEmpty } from '../../utils/predicates';
 
 import { Graph } from './graph';
